moved mod_info
