+++ src/commands/moderator_info/failToAcceptRatio.ts
