+++ src/commands/moderator_info/rankDist.ts
