+++ src/commands/moderator_info/scoreDist.ts
